// src/services/spotify/types.ts
export interface ExternalIds {
  upc: string;
}

// export interface Artist extends SimplifiedArtist {
//  followers: Followers;
//  genres: string[];
//  images: Image[];
//  popularity: number;
// }
//
// export interface Artists {
//  artists: Artist[];
// }
//
// export interface FollowedArtists {
//  artists: Page<Artist>;
// }
//
// export interface Followers {
//  href: string | null;
//  total: number;
// }
//
export interface ExternalUrls {
  spotify: string;
}
// ...
export interface LinkedFrom {
  external_urls: ExternalUrls;
  href: string;
  id: string;
  type: string;
  uri: string;
}
// ...
export interface Restrictions {
  reason: string;
}

// export interface Tracks {
//  tracks: Track[];
// }
//
export interface SimplifiedArtist {
  external_urls: ExternalUrls;
  href: string;
  id: string;
  name: string;
  type: string;
  uri: string;
}

export interface SimplifiedTrack {
  artists: SimplifiedArtist[];
  available_markets: string[];
  disc_number: number;
  duration_ms: number;
  episode: boolean;
  explicit: boolean;
  external_urls: ExternalUrls;
  href: string;
  id: string;
  is_local: boolean;
  is_playable?: boolean;
  linked_from?: LinkedFrom;
  name: string;
  preview_url: null | string;
  restrictions?: Restrictions;
  track: boolean;
  track_number: number;
  type: string;
  uri: string;
}

// export interface SavedTrack {
//  added_at: string;
//  track: Track;
// }
//
// export interface ExternalIds {
//  isrc: string;
//  ean: string;
//  upc: string;
// }
//
export interface Track extends SimplifiedTrack {
  album: SimplifiedAlbum;
  external_ids: ExternalIds;
  popularity: number;
}

interface AlbumBase {
  album_type: string;
  available_markets: string[];
  copyrights: Copyright[];
  external_ids: ExternalIds;
  external_urls: ExternalUrls;
  genres: string[];
  href: string;
  id: string;
  images: Image[];
  label: string;
  name: string;
  popularity: number;
  release_date: string;
  release_date_precision: string;
  restrictions?: Restrictions;
  total_tracks: number;
  type: string;
  uri: string;
}

interface SimplifiedAlbum extends AlbumBase {
  album_group: string;
  artists: SimplifiedArtist[];
}